requirements.txt
# === Core Data & ML ===
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.2
xgboost>=2.0.0
lightgbm>=4.0.0

# === Regime Detection ===
hmmlearn>=0.3.0

# === Technical Indicators ===
ta>=0.10.0

# === Sentiment ===
feedparser>=6.0.10
vaderSentiment>=3.3.2

# === OANDA Broker API ===
oandapyV20>=0.7.2

# === Environment & Config ===
python-dotenv>=1.0.0

# === Notifications ===
requests>=2.31.0

# === Visualization (Colab) ===
matplotlib>=3.7.0
seaborn>=0.12.0

# === Serialization ===
joblib>=1.3.0

# === Testing ===
pytest>=7.0.0
